Initial commit.
This is the first initial commit of the Chrome plugin "JustKeys". A
working connection between the currently displayed website and the
backend of the plugin is already in place. Also the JavaScript library
"Mousetrap" is used to inject a test keybinding into every visited
page and execute a test function when pressed.