Working link selection.
First working version of the link selection. After the follow link
function is activated ("f") and the visible links are highlighted one
can select the link to follow with the num keys. Backspace removes the
last typed number.

To hold the typed keys the object "Highlight" is introduced. It also
filters and counts the matching elements.